fix: enhance delete offer use case with additional error handling for unauthorized access and inactive farms, in order to pass all unit tests

+++ src/core/use-cases/delete-offer.ts
@@ -2,14 +2,17 @@
 import { OffersRepository } from "@/core/repositories/offers-repository";
 
 // Errors
-import { ResourceNotFoundError } from "@/core/errors/resource-not-found";
+import { FarmNotActiveError } from "@/core/errors/farm-not-active";
 import { ResourceClosedError } from "@/core/errors/resource-closed";
+import { ResourceNotFoundError } from "@/core/errors/resource-not-found";
+import { UnauthorizedError } from "@/core/errors/unauthorized";
 
 // Repositories
 import { CyclesRepository } from "@/core/repositories/cycles-repository";
 
 // Utils
 import { mostPast } from "@/core/utils/most-past";
+import { today } from "@/core/utils/today";
 
 interface DeleteOfferUseCaseRequest {
   farm_id: string;
@@ -31,7 +34,11 @@ export class DeleteOfferUseCase {
 
     const owner = offer.catalog?.farm?.id.equals(farm_id);
 
-    if (!owner) throw new ResourceNotFoundError("Oferta", offer_id);
+    if (!owner) throw new UnauthorizedError();
+
+    const active = offer.catalog?.farm?.status === "ACTIVE";
+
+    if (!active) throw new FarmNotActiveError();
 
     const cycle = await this.cyclesRepository.find("basic", {
       id: offer.catalog?.cycle_id.value,
@@ -42,6 +49,10 @@ export class DeleteOfferUseCase {
 
     if (offer.created_at < mostPast(cycle.order))
       throw new ResourceClosedError("Oferta", offer.id.value);
+
+    const isOfferingDay = cycle.offer.includes(today());
+
+    if (!isOfferingDay) throw new ResourceClosedError("Ciclo", cycle.id.value);
 
     await this.offersRepository.delete(offer);
   }
